requirements.txt
Keras
opencv-python
Pillow